requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
base58==2.1.1
bitarray==1.2.2
cachetools==4.2.4
certifi==2021.10.8
cytoolz==0.11.2
eth-abi==2.1.1
eth-account==0.5.6
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
frozenlist==1.2.0
google-api-core==2.3.2
google-auth==2.3.3
google-cloud-bigquery==2.31.0
google-cloud-core==2.2.1
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-status==1.43.0
hexbytes==0.2.2
idna==3.3
importlib-metadata==4.10.0
ipfshttpclient==0.8.0a2
jsonschema==3.2.0
lru-dict==1.1.7
multiaddr==0.0.9
multidict==5.2.0
netaddr==0.8.0
numpy==1.21.5
packaging==21.3
pandas==1.3.5
parsimonious==0.8.1
proto-plus==1.19.8
protobuf==3.19.1
psycopg2
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.12.0
pyparsing==3.0.6
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
rlp==2.0.1
rsa==4.8
six==1.16.0
toolz==0.11.2
typing-extensions==3.10.0.2
varint==1.0.2
web3==5.25.0
websockets==9.1
yarl==1.7.2
zipp==3.7.0

# Snow Flake
asn1crypto==1.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
cryptography==36.0.1
idna==3.3
oscrypto==1.2.1
pycparser==2.21
pycryptodomex==3.14.0
PyJWT==2.3.0
pyOpenSSL==21.0.0
pytz==2021.3
requests==2.27.1
six==1.16.0
urllib3==1.26.8
snowflake-connector-python==2.7.4
